Fix 'edit' command to allow new files
Previously, when one wanted to create mentorship notes for an exercise
that didn't have one file for, the `xr` function would reject this case.
But editors obviously can create files, so now the function lets one use
the 'edit' sub-command in order to create new notes, but still rejects
doing so when not in a exercise's directory and that no exercise name
has been explicitely given through a second argument.

Signed-off-by: Paul Mabileau <paulmabileau@hotmail.fr>