Isolate track-specific functions to separate script
This is the first step in supporting other tracks: move the _run_tests
and _run_benches functions to a separate script file in order to isolate
track-specific operations in dedicated modules. As of now, a new track
would just need to implement these two functions for the plugin to work
successfully with it.

Signed-off-by: Paul Mabileau <paulmabileau@hotmail.fr>